test: proper test-script filenames; make CI manual-only
File (script) names now state what each targets:
- tests/gui/test_headless.py        -> tests/gui/test_gui.py
- tests/test_cli_get_config.py      -> tests/test_get_config.py
- tests/realdata/test_plotting_gqi.py -> tests/realdata/test_plotting.py,
  with the GQI attempt-table test split out into tests/realdata/test_gqi.py.

CI trigger
- The workflow is now manual-only (workflow_dispatch); it no longer runs
  automatically on push or pull_request, to conserve GitHub Actions minutes.
  Trigger it on demand from the Actions tab ("Run workflow") or with
  `gh workflow run tests.yml`. Uncomment push / pull_request in tests.yml to
  re-enable automatic runs.
